feat: include `py3` statistics in `cp3`/`pp3` statistics
As was done for abi3, we now include py3 in cp3* / pp3* stats.

With abi3, we used the python tag to choose the minimum cp3 version and include it in more recent versions only.

With py3, we now use the `requires_python` metadata to choose the minimum cp3/pp3 version